+++ lib/expand_widget.dart
@@ -1,5 +1,6 @@
 library expand_widget;
 
 export 'src/expand_arrow.dart';
+export 'src/indicator_builder.dart';
 export 'src/expand_child.dart';
 export 'src/expand_text.dart';