+++ lib/src/indicator_builder.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/widgets.dart';
+
+/// This function is used to override the [ExpandArrow] widget for controlling
+/// a [ExpandChild] or [ExpandText] widget.
+typedef IndicatorBuilder = Widget Function(
+  BuildContext,
+  VoidCallback,
+  bool,
+);